+++ frontend/src/pages/search/index.tsx
@@ -8,12 +8,15 @@ import Header from "./components/Header";
 
 export default function index() {
   const [ifGuideClicked, setIfGuideClicked] = useState(false);
+
   const [date, setDate] = useState("");
   const [startHour, setStartHour] = useState("");
   const [endHour, setEndHour] = useState("");
   const [location, setLocation] = useState("");
   const [isScrolled, setIsScrolled] = useState(false);
-  const [locationDisplay, setLocationDisplay] = useState(true);
+
+  const [modalDisplay, setModalDisplay] = useState(true);
+  const [clickLoc, setClickLoc] = useState("");
 
   const handleGuidelineClick = () => {
     setIfGuideClicked(!ifGuideClicked);
@@ -29,17 +32,22 @@ export default function index() {
   };
 
   function handleClearBtn() {
+    setModalDisplay(true);
     setLocation("");
   }
 
   function handleSelectLocation(locationName?: string) {
     if (locationName) {
       setLocation(locationName);
-      setLocationDisplay(false);
+      setModalDisplay(false);
+      setClickLoc(locationName);
     }
   }
 
-  console.log(location);
+  function handleLocModify() {
+    setClickLoc("");
+    setModalDisplay(true);
+  }
 
   return (
     <S.Container>
@@ -53,7 +61,9 @@ export default function index() {
         isScrolled={isScrolled}
         handleClearBtn={handleClearBtn}
         handleSelectLocation={handleSelectLocation}
-        locationDisplay={locationDisplay}
+        modalDisplay={modalDisplay}
+        handleLocModify={handleLocModify}
+        clickLoc={clickLoc}
       />
       <Guide handleGuidelineClick={handleGuidelineClick} ifGuideClicked={ifGuideClicked} />
       {ifGuideClicked && <NoticeText />}
