+++ frontend/src/pages/upload/components/InputField.tsx
@@ -1,6 +1,8 @@
 import { useState } from "react";
+import { PencilIc } from "@assets/index";
 import * as U from "../styles";
 import styled from "styled-components";
+import FilteredModal from "./FilteredModal";
 
 interface Props {
   lostImgUrl: File;
@@ -9,6 +11,10 @@ interface Props {
 
 export default function InputField(props: Props) {
   const { lostImgUrl, handleSubmit } = props;
+  const [locModalDisplay, setLocModalDisplay] = useState(true);
+  const [strgModalDisplay, setStrgModalDisplay] = useState(true);
+  const [clickLoc, setClickLoc] = useState("");
+  const [clickStrg, setClickStrg] = useState("");
 
   const [date, setDate] = useState("");
   const [time, setTime] = useState("");
@@ -17,6 +23,42 @@ export default function InputField(props: Props) {
   const [location, setLocation] = useState("");
   const [storage, setStorage] = useState("");
   const [notes, setNotes] = useState("");
+
+  function handleClearLocationBtn() {
+    setLocation("");
+    setLocModalDisplay(true);
+  }
+
+  function handleClearStorageBtn() {
+    setStorage("");
+    setStrgModalDisplay(true);
+  }
+
+  function handleSelectLocation(locationName?: string) {
+    if (locationName) {
+      setLocation(locationName);
+      setClickLoc(locationName);
+      setLocModalDisplay(false);
+    }
+  }
+
+  function handleSelectStrg(storageName?: string) {
+    if (storageName) {
+      setStorage(storageName);
+      setClickStrg(storageName);
+      setStrgModalDisplay(false);
+    }
+  }
+
+  function handleLocModify() {
+    setClickLoc("");
+    setLocModalDisplay(true);
+  }
+
+  function handleStrgModify() {
+    setClickStrg("");
+    setStrgModalDisplay(true);
+  }
 
   return (
     <Container>
@@ -65,21 +107,47 @@ export default function InputField(props: Props) {
         <U.NotiText>아래 필수 항목들을 입력란에 기재해주세요.</U.NotiText>
         <U.FormGroup>
           <U.Label>습득 장소</U.Label>
-          <U.Input
-            type="text"
-            placeholder="물품을 습득한 장소를 입력해주세요."
-            value={location}
-            onChange={(e) => setLocation(e.target.value)}
-          />
+          {clickLoc ? (
+            <SearchResult>
+              <Text>{location}</Text>
+              <ModifyBtn onClick={handleLocModify}>
+                <PencilIcon />
+              </ModifyBtn>
+            </SearchResult>
+          ) : (
+            <>
+              <U.Input
+                type="text"
+                placeholder="물품을 습득한 장소를 입력해주세요."
+                value={location}
+                onChange={(e) => setLocation(e.target.value)}
+              />
+              {location && <ClearBtn onClick={handleClearLocationBtn}>X</ClearBtn>}
+            </>
+          )}
+          <FilteredModal input={location} modalDisplay={locModalDisplay} handleSelection={handleSelectLocation} />
         </U.FormGroup>
         <U.FormGroup>
           <U.Label>보관 장소</U.Label>
-          <U.Input
-            type="text"
-            placeholder="물품이 보관된 장소를 입력해주세요."
-            value={storage}
-            onChange={(e) => setStorage(e.target.value)}
-          />
+          {clickStrg ? (
+            <SearchResult>
+              <Text>{storage}</Text>
+              <ModifyBtn onClick={handleStrgModify}>
+                <PencilIcon />
+              </ModifyBtn>
+            </SearchResult>
+          ) : (
+            <>
+              <U.Input
+                type="text"
+                placeholder="물품이 보관된 장소를 입력해주세요."
+                value={storage}
+                onChange={(e) => setStorage(e.target.value)}
+              />
+              {storage && <ClearBtn onClick={handleClearStorageBtn}>X</ClearBtn>}
+            </>
+          )}
+          <FilteredModal input={storage} modalDisplay={strgModalDisplay} handleSelection={handleSelectStrg} />
         </U.FormGroup>
         <U.Divider />
         <U.OptionText>아래 선택 항목들을 입력란에 기재해주세요.</U.OptionText>
@@ -120,4 +188,57 @@ const Image = styled.img`
   border-radius: 12px;
   object-fit: cover;
   object-position: center;
+`;
+
+const ClearBtn = styled.button`
+  position: absolute;
+  top: 60%;
+  right: 10px;
+  padding: 2px;
+  border: none;
+  background: none;
+  color: #197a3a;
+  font-size: 1.4rem;
+  cursor: pointer;
+  transform: translateY(-50%);
+
+  &:hover {
+    color: #0056b3;
+  }
+`;
+
+const SearchResult = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  padding: 1.8rem;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  background-color: #ffe066;
+  color: #333;
+  font-size: 1.4rem;
+`;
+
+const ModifyBtn = styled.button`
+  border: none;
+  background: none;
+  color: #007bff;
+  font-size: 1.1rem;
+  cursor: pointer;
+
+  &:hover {
+    text-decoration: underline;
+  }
+`;
+
+const PencilIcon = styled(PencilIc)`
+  width: 1.5rem;
+  height: 1.5rem;
+`;
+
+const Text = styled.p`
+  color: #333;
+  font-size: 1.4rem;
+  font-weight: 500;
 `;
